Add full author

+++ src/main/java/com/fpt/controller/admin/PublisherController.java
@@ -1,12 +1,19 @@
 package com.fpt.controller.admin;
 
 import com.fpt.config.ProjectConfig;
+import com.fpt.entity.Author;
 import com.fpt.entity.Book;
 import com.fpt.entity.Publisher;
 import com.fpt.service.admin.PublisherServiceImpl;
+import com.fpt.specification.AuthorSpecification;
+import com.fpt.specification.PublisherSpecification;
+import com.fpt.specification.SearchCriteria;
+import com.google.gson.Gson;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
@@ -14,6 +21,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.Set;
 
@@ -25,10 +33,19 @@ public class PublisherController {
 
     @RequestMapping(method = RequestMethod.GET)
     public String list(
+            @RequestParam(name = "keyword", required = false) String keyword,
             @RequestParam(name = "page", defaultValue = "1") int page,
             @RequestParam(name = "limit", defaultValue = "5") int limit,
             Model model) {
-        Page<Publisher> publisherPage = publisherService.findAll(PageRequest.of(page - 1, limit));
+        Specification specification = Specification.where(null);
+        if (keyword != null && keyword.length() > 0) {
+            specification = specification
+                    .and(new PublisherSpecification(new SearchCriteria("name", "=", keyword)))
+                    .or(new PublisherSpecification(new SearchCriteria("description", "=", keyword)));
+        }
+        Page<Publisher> publisherPage = publisherService.findAllActive(specification,PageRequest.of(page - 1, limit));
+        model.addAttribute("keyword", keyword);
+
         model.addAttribute("publishers", publisherPage.getContent());
         model.addAttribute("currentPage", publisherPage.getPageable().getPageNumber() + 1);
         model.addAttribute("limit", publisherPage.getPageable().getPageSize());
@@ -74,11 +91,17 @@ public class PublisherController {
         return "admin/publisher/list";
     }
 
-    @RequestMapping(method = RequestMethod.POST, value = "/delete/{id}")
-    public String delete(@PathVariable(value = "id", required = false) long id, RedirectAttributes redirectAttributes) {
+    @RequestMapping(method = RequestMethod.DELETE, value = "/delete/{id}")
+    @ResponseBody
+    public String delete(@PathVariable(value = "id", required = false) long id, HttpServletResponse response) {
+        Publisher publisher = publisherService.getById(id);
+        if (publisher == null) {
+            response.setStatus(HttpStatus.NOT_FOUND.value());
+            return new Gson().toJson("Error");
+        }
         publisherService.delete(id);
-        redirectAttributes.addFlashAttribute("Success!", "Deleted contact successfully!");
-        return "redirect:" + ProjectConfig.PREFIX_ADMIN + ProjectConfig.PREFIX_ADMIN_PUBLISHERS;
+        response.setStatus(HttpStatus.OK.value());
+        return new Gson().toJson("Ok");
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/update/{id}")
